fix(channels): key entity chips, extract member picker, add reveal-on-add test
- Give each _EntityRow a stable ValueKey (role/member id, or a fixed key for
  "+ Add Member") so Flutter's element diffing doesn't reuse state across
  logically different rows when the member list grows — a latent bug the
  keyless list comprehension was exposed to once members could be added
  in place.
- Extract _MemberPickerDialog into its own file
  (channel_member_picker_dialog.dart) as ChannelMemberPickerDialog, bringing
  channel_permissions.dart back under the project's 1000-line guideline
  (it landed at 1021 lines after the responsive-layout changes).
- Add a regression test for the "selecting a chip (including a freshly added
  member) scrolls it into view" behavior described in the PR: the existing
  suite only exercised pre-existing chips via manual `tester.ensureVisible`
  calls, so the actual auto-reveal-on-select codepath for a brand new member
  overwrite had no coverage.
